simd: improve estimated signature count
The command line parameter can easily be full of lies, or cause us to
overallocate.  This can be deadly when dealing with huge numbers of
hashes.  Instead, we'll count the number of lines in the input file, and
use that to get a better estimate on how big our tables should be.

This also fixes a bug when in sharding mode we wouldn't reduce the
number of hashes assigned to this machine, and instead preallocate for
the entire batch (thus mostly eliminating the benefit of sharding for
reduced memory usage.)